refactor(tests): extract some factory methods to trait
Simply to make it easier to customize factory classes in app-level
