+++ database/factories/BusinessFactory.php
@@ -2,10 +2,8 @@
 
 namespace Database\Factories;
 
-use Creasi\Base\Models\Address;
 use Creasi\Base\Models\Business;
-use Creasi\Base\Models\Enums\FileUploadType;
-use Creasi\Base\Models\FileUpload;
+use Creasi\Base\Models\Concerns\Factories;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
 /**
@@ -13,6 +11,9 @@ use Illuminate\Database\Eloquent\Factories\Factory;
  */
 class BusinessFactory extends Factory
 {
+    use Factories\WithAddress;
+    use Factories\WithFileUpload;
+
     protected $model = Business::class;
 
     /**
@@ -26,17 +27,5 @@ class BusinessFactory extends Factory
             'phone' => '08'.$this->faker->numerify('##########'),
             'summary' => $this->faker->sentence(4),
         ];
-    }
-
-    public function withAddress(): static
-    {
-        return $this->has(Address::factory());
-    }
-
-    public function withFileUpload(FileUploadType $type = null): static
-    {
-        return $this->hasAttached(FileUpload::factory(), [
-            'type' => $type ?? $this->faker->randomElement(FileUploadType::cases()),
-        ], 'files');
     }
 }
