+++ database/factories/PersonnelFactory.php
@@ -2,16 +2,9 @@
 
 namespace Database\Factories;
 
-use Creasi\Base\Models\Address;
-use Creasi\Base\Models\Business;
-use Creasi\Base\Models\Enums\EmploymentStatus;
-use Creasi\Base\Models\Enums\EmploymentType;
-use Creasi\Base\Models\Enums\FileUploadType;
+use Creasi\Base\Models\Concerns\Factories;
 use Creasi\Base\Models\Enums\Gender;
-use Creasi\Base\Models\FileUpload;
 use Creasi\Base\Models\Personnel;
-use Creasi\Base\Models\Profile;
-use DateTimeInterface;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
 /**
@@ -19,6 +12,10 @@ use Illuminate\Database\Eloquent\Factories\Factory;
  */
 class PersonnelFactory extends Factory
 {
+    use Factories\AsPersonnel;
+    use Factories\WithAddress;
+    use Factories\WithFileUpload;
+
     protected $model = Personnel::class;
 
     /**
@@ -43,43 +40,5 @@ class PersonnelFactory extends Factory
         return $this->state([
             'user_id' => null,
         ]);
-    }
-
-    public function withProfile(Gender $gender = null): static
-    {
-        return $this->has(Profile::factory(), 'profile')->state(fn () => [
-            'name' => $this->faker->firstName($gender?->toFaker()),
-        ]);
-    }
-
-    public function withCompany(
-        bool $primary = null,
-        EmploymentType $type = null,
-        EmploymentStatus $status = null,
-        false|DateTimeInterface $startDate = null,
-    ): static {
-        if (null === $startDate) {
-            $startDate = $this->faker->dateTime();
-        }
-
-        return $this->hasAttached(Business::factory(), [
-            'is_primary' => $primary,
-            'type' => $type ?? $this->faker->randomElement(EmploymentType::cases()),
-            'status' => $status ?? $this->faker->randomElement(EmploymentStatus::cases()),
-            'start_date' => $startDate?->format('Y-m-d'),
-            'finish_date' => null,
-        ], 'employers');
-    }
-
-    public function withAddress(): static
-    {
-        return $this->has(Address::factory());
-    }
-
-    public function withFileUpload(FileUploadType $type = null): static
-    {
-        return $this->hasAttached(FileUpload::factory(), [
-            'type' => $type ?? $this->faker->randomElement(FileUploadType::cases()),
-        ], 'files');
     }
 }
