+++ database/factories/ProfileFactory.php
@@ -2,9 +2,7 @@
 
 namespace Database\Factories;
 
-use Creasi\Base\Models\Enums\Education;
-use Creasi\Base\Models\Enums\Religion;
-use Creasi\Base\Models\Enums\TaxStatus;
+use Creasi\Base\Models\Enums;
 use Creasi\Base\Models\Profile;
 use Creasi\Nusa\Models\Regency;
 use Illuminate\Contracts\Database\Query\Builder;
@@ -30,9 +28,9 @@ class ProfileFactory extends Factory
             'nik' => $nik = $this->faker->nik(null, $birthDate = $this->faker->dateTime()),
             'birth_date' => $birthDate->format('Y-m-d'),
             'birth_place_code' => $birthPlace->code,
-            'education' => $this->faker->randomElement(Education::cases()),
-            'religion' => $this->faker->randomElement(Religion::cases()),
-            'tax_status' => $this->faker->randomElement(TaxStatus::cases()),
+            'education' => $this->faker->randomElement(Enums\Education::cases()),
+            'religion' => $this->faker->randomElement(Enums\Religion::cases()),
+            'tax_status' => $this->faker->randomElement(Enums\TaxStatus::cases()),
             'tax_id' => $nik,
         ];
     }
