+++ src/Models/Concerns/Factories/WithFileUpload.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Creasi\Base\Models\Concerns\Factories;
+
+use Creasi\Base\Models\Enums\FileUploadType;
+use Creasi\Base\Models\FileUpload;
+
+/**
+ * @mixin \Illuminate\Database\Eloquent\Factories\Factory
+ */
+trait WithFileUpload
+{
+    public function withFileUpload(FileUploadType $type = null): static
+    {
+        return $this->hasAttached(FileUpload::factory(), [
+            'type' => $type ?? $this->faker->randomElement(FileUploadType::cases()),
+        ], 'files');
+    }
+}
